A fu*king  stupid bug
I have added some few features with a bug. It happens when the
destructor is called, but I did not find something wrong with my code.
and I still think my code is nothing wrong logically and correct
grammer. And I spent at least 2 hours to fix it. But I do not find it,
so I ask my friend to recode it and there is not bad thing happened in
his Mac. Maybe I should change IDE from Visual Stdio 2015 to another or
change the complior from vc++ to clang, but I want to try it tomorrow,
CUZ I have suffered from it enough and I am REALLY REALLY tired. So let
me restart it tomorrow.