Forget to fix some left bugs and modify some details forgotten to comment, but get a compilor warning. :(
Fix the left bugs:
1. MySTLString::const_reference MySTLString::front() const
2. const MySTLString::CharT* MySTLString::data() const

Comment the code for test.
But get a compilor warning C4172 returning address of local variable or
temporary
A function returns the address of a local variable or temporary
object.Local variables and temporary objects are destroyed when a
function returns, so the address returned is not valid.
I think maybe I should redesign those functions who have the warning. :(