Have added the implement of class MyIterator and its derived class, but I get a bad feeling to it.
I am really worry about its implement and I am afraid that I have to
debug it in the next two or three days.
And after finish it, I would like to transfrom the normal class to the
class template.
I will test it in tomorrow, hope it will not make me overwhelmed.